Ensure tests reflect the moved utility methods
